cscope-directory property now read from (point)
This property is added to the whole block of text in a particular result, so
there was no reason to move the point before querying the property. An advantage
in keeping the (point) is that at the end-of-buffer, the property is absent,
thus the current directory will be picked up. Thus at the end-of-buffer the user
can do 'M-x cd', and then a cscope search will be done in that directory