cscope-indexer integrated into the .el. Customizable directories/files to index
I no longer need an external cscope-indexer script. This "script" is now
constructed on the fly in the emacs lisp, and executed as a cmdline argument by
the shell.

Furthermore, the list of directories to ignore is now customizable in emacs, as
is the list of file extensions to index