Use `read-string' rather than `read-from-minibuffer'
This applies to functions prompting for symbols, filenames, etc. to enable
better behaviour and input history.

The main benefit is if the symbol under point is _not_ the one that I want to
search for when I run `cscope-find-global-definition’ (or any of the similar
bindings). Before this patch I have to delete the symbol that was chosen, then
enter my own. After the patch I can simply enter my preferred symbol without the
need to delete the automatically chosen symbol. The ability to just hit ‘ret’ to
get the default symbol is retained.