fixed possible infinite loop with TRAMP
If running cscope over TRAMP without first creating an index, xcscope.el would
infinite loop looking for the top-level of the directory hierarchy. It was
explicitly looking for a string representation of the top level. Now I exit when
the parent directory string is exactly the same as the current directory string