Removed 'cscope-directory property. Parsing visible text instead
Previously I was using a property to mark the directory of a cscope search in
the *cscope* buffer. This was not working when `cscope-database-regexps' was
non-null, making possible multiple directories in a single search. Since the
visible text states the directory of each individual search, I now simply parse
that out to get the directory