Fix "<file> is not readable or exists" when using xcscope over TRAMP.
1. The "<file> is not readable or exists" issue:

When using xcscope over TRAMP, the database can be found and the query
results shown correctly.  However, the results cannot be jumped to
with a message "<file> is not readable or exists".  It seems that
xcscope tries the filename on my local machine.

2. Possible fix (this commit):

When adding text properties, prepend the filename with remote file
identifier if 'default-directory' is remote.