Quote and escape file names in index file and allow whitespaces
cscope can handle files with whitespaces and other special characters,
but they must be properly quoted and escaped in the index file. This
can be fixed by filtering the output of 'find' through 'sed'.

cscope still complains when a symbol is found in a file with a
whitespace in its name ("File does not have expected format") but at
least it works (partially) now. On the other hand, filenames with
double quotes or backslashes seem to work just fine.

In addition to that, the regular expression that parses the cscope
output needs to be modified to allow whitespaces in file names. The
new regular expression still does not allow tabs in file names, but
they are extremely unlikely to happen and cscope does not seem to
support them anyway.